+++ core/src/main/java/com/orientechnologies/orient/core/storage/cluster/v0/OClusterPositionMapV0.java
@@ -36,6 +36,7 @@ import java.util.Arrays;
  * @since 10/7/13
  */
 public class OClusterPositionMapV0 extends OClusterPositionMap {
+
   private long fileId;
 
   OClusterPositionMapV0(
@@ -97,7 +98,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
       }
 
       final long index = bucket.add(pageIndex, recordPosition);
-      return index + cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
+      return index + (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
     } finally {
       cacheEntry.close();
     }
@@ -131,7 +132,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
       }
 
       final long index = bucket.allocate();
-      return index + cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
+      return index + (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
     } finally {
       cacheEntry.close();
     }
@@ -231,7 +232,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
         } else {
           int entriesCount = 0;
           final long startIndex =
-              cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
+              (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
 
           result = new OClusterPositionEntry[resultSize];
           for (int i = 0; i < resultSize; i++) {
@@ -292,7 +293,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
         } else {
           int entriesCount = 0;
           final long startIndex =
-              cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
+              (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
 
           result = new long[resultSize];
           for (int i = 0; i < resultSize; i++) {
@@ -362,7 +363,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
         int entriesCount = 0;
 
         final long startPosition =
-            cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
+            (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
         result = new long[resultSize];
 
         for (int i = 0; i < resultSize; i++) {
@@ -442,7 +443,9 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
     return ORID.CLUSTER_POS_INVALID;
   }
 
-  /** Returns the next position available. */
+  /**
+   * Returns the next position available.
+   */
   long getNextPosition(final OAtomicOperation atomicOperation) throws IOException {
     final long filledUpTo = getFilledUpTo(atomicOperation, fileId);
 
@@ -460,6 +463,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
   }
 
   public static class OClusterPositionEntry {
+
     private final long position;
     private final long page;
     private final int offset;
