+++ core/src/main/java/com/orientechnologies/orient/core/storage/index/nkbtree/normalizers/DecimalKeyNormalizer.java
@@ -8,6 +8,7 @@ import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 
 public class DecimalKeyNormalizer implements KeyNormalizers {
+
   private static final BigInteger BIG_INT_TEN = new BigInteger("10");
   private static final BigInteger BIG_INT_ONE = new BigInteger("1");
   private static final BigInteger BIG_INT_ZERO = new BigInteger("0");
@@ -43,7 +44,7 @@ public class DecimalKeyNormalizer implements KeyNormalizers {
 
   private BigDecimal scaleToDecimal128(final BigDecimal rawValue) {
     final BigDecimal value = clampAndRound(rawValue);
-    long exponent = (long) (-value.scale());
+    long exponent = -value.scale();
     if (exponent >= -6176L && exponent <= 6111L) {
       if (value.unscaledValue().bitLength() > 113) {
         throw new AssertionError(
